Move eliminate_assume() to sympy/assumptions.ask.py
This function is only used in ask().
